+++ torch_frame/transforms/__init__.py
@@ -1,0 +1,11 @@
+from .base_transform import BaseTransform
+from .fittable_base_transform import FittableBaseTransform
+from .categorical_catboost_encoder import CategoricalCatBoostEncoder
+from .mutual_information_sort import MutualInformationSort
+
+__all__ = functions = [
+    'BaseTransform',
+    'CategoricalCatBoostEncoder',
+    'FittableBaseTransform',
+    'MutualInformationSort',
+]